+++ ShowdownBot/Console.Help.cs
@@ -6,7 +6,7 @@ using static ShowdownBot.Global;
 using static ShowdownBot.GlobalConstants;
 namespace ShowdownBot
 {
-    //Todo: Move this to an external (XML?) file?
+    
 
     /// <summary>
     /// Contains the in-console help documentation for commands.